Fix: Send関数がsenderに記述されていたのでhubに書き直しました
