+++ internal/broadcaster/sender.go
@@ -2,7 +2,6 @@ package broadcaster
 
 import (
 	"encoding/json"
-	"fmt"
 	"log"
 
 	"github.com/shii-park/Stshoot_Backend/internal/model"
@@ -13,15 +12,6 @@ import (
 type SenderClient struct {
 	Hub  *Hub
 	Conn *websocket.Conn
-}
-
-func (r *Receiver) Send(message []byte) error {
-	r.Mu.RLock()
-	defer r.Mu.RUnlock()
-	if r.Conn == nil {
-		return fmt.Errorf("no receiver client connected")
-	}
-	return r.Conn.WriteMessage(websocket.TextMessage, message)
 }
 
 func (c *SenderClient) ReadPump() {
